refactor: 입력값에 따라 프로그램 생성

+++ src/main/java/domain/ui/UserInterface.java
@@ -1,5 +1,5 @@
 package domain.ui;
 
 public interface UserInterface {
-    int runProgram();
+    double runProgram();
 }
